+++ packages/bluesky/test/src/services/feed_service_test.dart
@@ -5,7 +5,6 @@
 // 📦 Package imports:
 import 'package:atproto/atproto.dart' as atp;
 import 'package:atproto_core/atproto_core.dart';
-import 'package:bluesky/bluesky.dart';
 
 // 🌎 Project imports:
 import 'package:bluesky/src/ids.g.dart';
@@ -14,6 +13,7 @@ import 'package:bluesky/src/services/entities/feed.dart';
 import 'package:bluesky/src/services/entities/feed_generator.dart';
 import 'package:bluesky/src/services/entities/feed_generator_info.dart';
 import 'package:bluesky/src/services/entities/feed_generators.dart';
+import 'package:bluesky/src/services/entities/feed_interaction.dart';
 import 'package:bluesky/src/services/entities/likes.dart';
 import 'package:bluesky/src/services/entities/post_thread.dart';
 import 'package:bluesky/src/services/entities/posts.dart';